Temporary fix for the admin bar issue
Sometimes Selenium cannot click an element because the "Element is
not clickable at point". This can be because the target element is
underneath the admin bar.

This is a quick fix for that issue, by forcibly setting the z-index
of the admin bar so that it appears beneath the target element.

Ref #1.