Use getHtml() and strip out tags instead of getText()
When the admin menu is collapsed getText() returns an empty string for all items except Comments and Plugins. To get around this we obtain the HTML and strip out any HTML tags and their contents.
Also, when the menu is collapsed we cannot see submenu items to click them unless we focus on the parent menu item.

Fixes #2.